Versie 0.8.8 - KVK-nummer naar bigInt, filenames naar nullable

+++ database/migrations/create_kvk_numbers_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('kvk_numbers', function (Blueprint $table) {
             $table->id();
             $table->foreignId('company_id');
-            $table->integer('kvk');
+            $table->bigInteger('kvk');
             $table->timestamps();
             $table->unique(['company_id', 'kvk']);
         });
